Add a result converter to the ListenableFuture adapter
Very often in the client implementation, we need to apply a simple conversion to the driver result before invoking the handler.

Without the converter param, we had to create an extra handler for the async result, and invoke the user handler manually.

This commit will simplify the client implementation code and avoid creating extra handlers/futures (example given in ResultSet implementation)